Refactor location of options file.
