+++ src/cplus_plugin/gui/settings/report_options.py
@@ -1,0 +1,3 @@
+# coding=utf-8
+
+"""Plugin report settings."""
