+++ src/cplus_plugin/ui/cplus_settings.ui
